Add Transactions and create method

+++ db/migrate/20220523153402_create_entities.rb
@@ -2,7 +2,7 @@ class CreateEntities < ActiveRecord::Migration[7.0]
   def change
     create_table :payments do |t|
       t.string :name
-      t.money :amount
+      t.float :amount
       t.references :author, index: true, foreign_key: { to_table: :users }
 
       t.timestamps
